backend/requirements_tflite.txt
 # Requirements cho Raspberry Pi với TFLite + Camera Server
flask==3.0.0
flask-cors==4.0.0
cloudinary==1.36.0
pymongo==4.6.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
python-dotenv==1.0.0
requests==2.31.0

# TFLite runtime cho Raspberry Pi (cài riêng)
# pip install --extra-index-url https://google-coral.github.io/py-repo/ tflite_runtime
# Hoặc nếu không có tflite-runtime, dùng tensorflow-lite (nhẹ hơn tensorflow đầy đủ)
# tensorflow==2.15.0

